mmalvideodec: Initial outline implementation of h.264 decoder.
This commit adds a rough initial implementation of a h.264 decoder.
It is far from complete, with many things stil to be implemented.
Still, it can decode frames and is a useful starting point.

The majority of the implementation is in the `GstMMALVideoDec`
baseclass, and additional format support, (such as mpeg2), will
require their own elements.

Some outstanding tasks:

- Neither flush() nor drain() are implemented.

- Port "event" handling.

- Dynamic format reconfiguration (for adaptive streams).

- Seek & discontinuity.

- No support for opaque buffers yet.

- change_state() needs proper implementation.

- Detection and signalling of interlace mode.

- Should not need to copy input Gst buffers to MMAL buffers.

- Support for opaque buffers.

- Much tidying required!