mmalvideodec: Implement drain() and use it in set_format().
When the `GstVideoDecoder` baseclass calls the `set_format()` vmethod to
inform us of new caps, we need to reconfigure both the input and output
ports, resize the buffers in the buffer pool, and notify downstream of the
caps change.  Downstream notification is done via the `GstVideoDecoder`
`gst_video_decoder_set_output_state()` function, which will cause the new
caps to be set downstream when we next push an output frame by calling
`gst_video_decoder_finish_frame()`.

Now the issue is that there may still be output frames pending (with the
"old" format) when `set_format()` is called, so we don't want to tell
downstream the new format before they have been processed and sent
downstream.  Reconfiguring the output buffer pool before then would also
likely give us trouble.  So, we need to first drain the decoder and wait
for all pending frames before changing anything.  Hence we need the
`drain()` function.

We implement `drain()` by sending an (empty) MMAL buffer to the decoder
input port with the EOS flag set on it.  We then wait for a matching EOS
output buffer to be pushed out of the decoder output port.  Once we see
this, we know that there are no more pending frames with the old format to
come.

Since we're waiting for the EOS output buffer, we need to take care to
keep returning empty output buffers back to the decoder as pending frames
in front of EOS are collected from the output port.  Otherwise, we can
easily get into the situation where we never see EOS because the decoder
does not have an output buffer (header) to put it in.

With these changes, the decoder is now able to keep continuously decoding
frames through format changes.