mmalvideosink: Support any input video size for plain buffers
As long as input video size is aligned with MMAL expectations (32 for width
and 16 for height), GStreamer buffer data can be passed directly to MMAL
without copying.  When video size is not aligned, raw frame data needs to
be copied into MMAL buffer payload allocated in that situation.

Example tests:

  * optimal case

    GST_DEBUG=mmal*:5,video-info:6 gst-launch-1.0 videotestsrc ! \
        video/x-raw,width=640,height=320 ! mmalvideosink

  * misaligned case

    GST_DEBUG=mmal*:5,video-info:6 gst-launch-1.0 videotestsrc ! \
        video/x-raw,width=551,height=390 ! mmalvideosink