mmalvideodec: Don't copy input GstBuffers to MMAL buffers.
This commit borrows the technique from mmalvideosink to avoid copying the
input buffers into MMAL buffers.  Instead, we let the MMAL buffers hold
a simple housekeeping structure that points to the GstBuffer and the
mapped memory.

Unfortunately, things are not as simple as they might be as it seems that
the decoder gets upset if it receives an input buffer that is larger than
the buffer size specified on the decoder input port.  I guess this is
probably because it will be copying from userspace into VC buffers, so it
will need to size those appropriately.

We need to take care of the (likely) case where the GstBuffer containing
the encoded frame is larger than the buffer size we configured on the
input port.

We solve this problem by mapping the GstBuffer memory just once, and
copying the mapping info to each MMAL buffer header we have to send.
Since we only want to unmap the GstBuffer once as well, we form a
linked-list structure over the `MappedBuffer` structures we attach to
each MMAL buffer header.  We know the last reference has been reached
when the `MappedBuffer` referenced by a freed MMAL buffer header has no
siblings.