mmalvideodec: Take src caps framerate from decoder output port.
It seems that we can't always rely on upstream elements to provide the
correct framerate via sink caps.

The decoder does a decent job of detecting the framerate (presumably by
looking at SPS header in H264 case), so we would like to use that instead.

This commit defers the call to update_src_caps(), such that, rather than
being called directly from set_format(), it is called once the first output
frame with the new format is received from the decoder.  This gives the
decoder a chance to parse the header and work out framerate.

Unfortunately, it seems that it is also necessary to commit the output
port format at this point, which requires first disabling the output port
again.  Otherwise the output port format is stale, and doesn't hold the
new framerate.  There might be an easier way to do this, but I haven't
found it yet.

We also set the framerate to `0` on the decoder input port, rather using
the (potentially bogus) framerate we received via sink pad caps as otherwise
the decoder seems to just take this hint as the truth.