mmalvideodec: Query decoder for interlace/progressive and set src caps and output buffer flags accordingly.
Any downstream deinterlace element needs to know whether the frames it is
receiving are interlaced, the interlace mode, field order, and whether the
first field is repeated.

Interlace vs progressive, and interlace mode can be communicated via src
caps, whereas field-order and field repeat need to be specified as flags
on the outgoing GstBuffers.

This commit is the first pass at implementing interlace detection and
signalling via caps and buffer flags.  The query is done by reading the
MMAL interlace parameter on the decoder component following a format change.

There is a small problem though in that, for mixed-mode interlace, it is
possible to have some frames progressive, whilst others are interlaced, so
really we want to be checking the interlace flags on each frame buffer that
comes out of the decoder.  Unfortunately though, the video flags on the
output MMAL buffer headers seems to be zero, even when the frame is
interlaced, so this is not much use.  It's possible that the flags are only
set in the case of mixed-mode though, so that will need to be tested with
some appropriate test material.

With these changes, inserting the existing `deinterlace` element between
`mmalh264dec` and `mmalvideosink` now actually deinterlaces the video.
More thorough testing is required, with different interlace modes, but
this is a start.