mmalvideosink: Support smooth resolution changes (fixed buffer size)
Turns out it's possible to use renderer component for smooth resolution
(format) transitions without disabling it and its port.  The goal it to
have the last frame displayed instead of having a black screen flash.

For this to happen, VC also holds userland buffer which is inconvenient
for resizing buffer pool.  To avoid resizing, a fixed size buffer pool is
allocated.  Currently the upper limit is 1080p resolution.