mmalvideodec: Support MMAL opaque output buffers.
This commit enables the decoder to negotiate the use of "opaque" MMAL
output buffers, if the downstream element supports it.

Opaque buffers avoid the need for the decoded frames to be mapped into
system memory and any copying of the buffers.  Note that MMAL also supports
zero-copy of buffers via shared-memory between VC & ARM, but that doesn't
net us any benefit when using opaque buffers since only the buffer headers
need to be exchanged, and these should fit in the control messages anyway.
For that reason, we don't enable zero-copy.

In order to transport MMAL buffer headers via GstBuffers, we use a custom
`GstMemory` type: `GstMMALOpaqueMemory`, which holds a pointer to the MMAL
buffer header.  Separate functions are provided for setting and getting
the MMAL buffer header in the memory object.

The `GstMMALOpaqueMemory` object takes a reference to the MMAL buffer header,
which must be released when the `GstBuffer` is returned to the pool. This is
achieved by the use of a custom buffer pool: `GstMMALOpaqueBufferPool`,
which the decoder itself creates in `decide_allocation()`.  This is the
only real reason we need the custom buffer pool.

The downstream element is expected to implement `propose_allocation()` and
signal that it wants to use opaque buffers by providing an
`GstMMALOpaqueAllocator`.  This specialised allocator type doesn't do a
great deal, other than signal to the decoder that opaque buffers should be
used.

Note: To force the use of plain buffers, even if downstream element supports
opaque, just insert a caps filter inbetween.  For example:

  mmalh264dec ! video/x-raw ! mmalvideosink

Or, to require opaque:

  mmalh264dec ! "video/x-raw(memory:MMALOpaque)" ! mmalvideosink

Default decoder behaviour (i.e. without filter) is to first attempt to
negotiate opaque, then fallback to plain buffers.