mmalvideosink: Support MMAL opaque input buffers.
This commit adds MMAL opaque buffer support to the video sink.

Key changes:

- video sink implements `propose_allocation()` and provides an instance of
`GstMMALOpaqueAllocator`.

- Following `set_caps()`, reconfiguration is deferred until the next buffer
is received from upstream.  This allows the sink to determine that upstream
has decided to use opaque buffers, rather than plain.

- The sink doesn't allocate buffer pool if when using opaque buffers.

- In opaque mode, `show_frame()` extracts the MMAL buffer header from the
`GstMemory` and increments it's reference count.  This reference will be
dropped by the port callback function, when the renderer component is
finished with the buffer.

- The `GstMemory` is also holding a reference to the MMAL buffer header,
which will be freed when it's parent `GstBuffer` is returned to it's buffer
pool.  This, together with the fact that `show_frame()` takes a ref each
time a frame is to be displayed, avoids a double-deref problem on pre-roll.