mmalvideodec:  Set MMAL_PARAMETER_EXTRA_BUFFERS on output port when using opaque buffers.
This parameter is required to configure additional "real" buffers on the
VideoCore side when using opaque buffers for output.  Without this, we will
see the decoder get into a state where it consumes more and more input
buffers, without producing any output frames, until we run out of input
buffers.  This may be due to the decoder requiring more references frames,
and/or length of the DPB, or something else.  We use the same value of 20
that VLC uses.  This seems a tad high (@~3MB per frame for 1080p), but only
more extensive testing will tell us.

We also allocate the same additional number of buffer headers in opaque
mode, but keep a separate value for when using plain buffers.