mmalvideosink: Provide support for video window size and position
mmalvideosink is a GStreamer element that can render raw video using MMAL
APIs.  By default it renders video in full-screen mode.  In some scenarios
it is desired to render scaled video into specific region on the screen at
given position and with given size.

This commit adds support for 'destinationwindow' property which allows to
specify destination video window region as a structure of doubles (0.0 to
1.0):

  x      - relative horizontal position (x * screen-width)

  y      - relative vertical position (y * screen-height)

  width  - relative width (width * screen-width)

  height - relative height (height * screen-height)

Values falling outsize of this range are clamped.  Also a minimum of the
final video size is guaranteed to be no less than 1/8 of the screen size
for sanity.

To test on the command line:

    gst-launch-1.0 videotestsrc ! video/x-raw,width=720,height=480 ! \
    mmalvideosink destinationwindow="destinationwindow, x=(double)0.2, y=(double)0.2, width=(double)0.3, height=(double)0.3"