mmalvideosink: Support switching between opaque & plain buffers after initial negotiation.
Usually, we expect that once caps have been negotiated and allocation
decided we will be using either opaque or plain buffers for the entire
stream.

However, in case of some pipelines, it is possible for buffer type to be
re-negotiated when there is a format switch in the stream.  In particular,
if stream switches from progressive to interlaced, and we happen to have
deinterlace element in-between decoder and sink, then deinterlace suddenly
becomes involved in allocation query, rather than just passing it through,
and prevents the use of MMAL opaque buffers.

In order to handle this situation, the sink needs to detect when buffer
type has changed, and ensure it reconfigures it's input port and creates,
or destroys the plain buffer pool, as appropriate.  This commit adds that.

The following should reproduce the buffer type switching:

  gst-play-1.0 --videosink mmalvideosink \
http://rdmedia.bbc.co.uk/dash/ondemand/testcard/1/client_manifest-events.mpd

(wait for 1080i representation change).