mmalvideodec: Simplify filling plain aligned buffers
mmalvideodec element operates in various modes when it comes to writing
output frames.  In one of them, plain buffers are used and data returned
from VC is copied over.  Specifically, when the frame size meets the VC
alignment, strides don't need to be taken into account and the operation is
a simply copy.

This commit simplifies code that handles aligned plain buffers by calling a
single GStreamer function.  Tested with 720p stream and plain buffers
enforced.