mmalvideodec: Disable MMAL component before destroying it
When playing a stream with pipeline where deinterlace component is
instantiated, reaching EOS triggers deinterlace component to be closed
first.  In that situation decoder crashes somewhere in VC function when
closing.  If interrupting playback from the command line (EOS sent as well)
the order is opposite in which case everything is fine.

This commit disable decoder component before destroying although this still
doesn't help.