mmalvideodec: Only destroy decoded frames pool after the port is disabled
Although the pool is destroyed after the port is flushed and all buffers
are returned, in general decoded frames pool should be destroyed only
after the output port is disabled.  Disabling (or flushing) the output
port invoked the output port callback which puts all returned buffers
onto that queue.  This could lead to use the queue after it was freed.

In practice, because the port is flushed in the first place, use after
free doesn't occur but for sanity proper ordering of object destruction
is applied.