mmalvideodec,deinterlace,sink: Enable zero-copy on ports when using opaque buffers.
Although it may appear so, this commit is change is not being made for
performance reasons, but rather to avoid a nasty segfault after EOS when
the deinterlacer is present in the pipeline.

The segfault happens when the decoder component is destroyed in
`gst_mmal_video_dec_close()`, but can be avoided if any one of the decoder,
image_fx (deinterlacer), or renderer components is *not* closed, or if the
decoder output pool is not destroyed.  Since those workarounds would likely
result in memory leaks, they are not good solutions for this mystery
problem.

It appears that enabling the zero-copy parameter on all ports using opaque
buffers (but particularly image_fx output port and renderer input port),
also avoids the issue.  Although it is unclear why this should be the case,
it seems that other implementations (e.g. VLC, XBMC) also enable zero-copy
when using opaque buffers, so this is likely just a better tested
configuration.

N.B. Zero-copy does *not* seem to offer any performance benefit when using
opaque buffers, since the bulk transfer is anyway avoidable because the
opaque handles should fit in the control messages between VC and ARM.