mmal: Make operations on MMAL buffer-header refcount thread-safe.
Buffers are exhanged via MMAL components indirectly by passing
"buffer headers" between component ports.  These headers are
structures which contain some metadata describing the buffer,
plus a pointer to the buffer payload.

Buffer headers themselves are allocated from a pool which is
associated with a port on the owning component.  The headers
are individually reference-counted and are released back to
the pool once the reference count reaches 0.

The client client calls `mmal_buffer_header_acquire()` to take
a reference to a header, and `mmal_buffer_header_release()`
to release it.  Other functions in the mmal library also call
these functions, as part of their implementation.

Unfortunately, these functions do not handle the header
reference-count in a thread-safe way, so in a multi-threaded
environment, (and particularly in SMP case), we can end-up
with corruption of the reference-count, which can result in
buffers being leaked (i.e. not returned to the pool).

We see this problem in our elements, since these functions
are called both from MMAL callbacks (on some "MMAL" thread),
and also from GST threads.  This ultimately leads to the
component ports being starved of buffers and the pipeline
grinding to a halt.

This commit adds a workaround for the thread-safety issue,
in the form of a re-implementation of the acquire/release
functions.

Due to the way dynamic linking works, our local implementation
overrides the implementation in the MMAL lib.

Our replacement implementation of these functions uses the
compiler-intrinsic atomic operations to manipulate the
refcount.

This fix is somewhat instrusive as it is dependent on some
the private MMAL implementation details, and it would be more
ideal if the issue were fixed upstream.