mmalvideosink: Do not submit same opaque buffer more than once
Sink elements are special as they handle an edge case: when the pipeline
goes into pausing state, they are supposed to show the first frame.
This is called pre-roll.

MMAL elements can exchange opaque buffers which is very efficient as it
avoids copying a lot of data.  These buffers themselves do not seem to
be handled in a thread-safe manner: VCHIQ component modifies them by
writing back some information from VC.  If at the same time GStreamer
thread tries to send such buffer to VC, this might cause a confusion as
inconsistent information about the buffer is queued for sending into VC.
This has been seen (and suspected to be the reason of) EIO errors
reported by VC.

Since buffers are sent twice subsequently only in pre-roll phase [1], a
simple way of avoiding this is to ensure that the same buffer is not
sent to the MMAL component subsequently.  This commit just ensures this.
No EIO errors where seen afterwards.

[1] http://cgit.freedesktop.org/gstreamer/gstreamer/tree/docs/design/part-preroll.txt:

    When going to PAUSED and PLAYING a buffer should be queued in the pad.
    We also make this requirement for going to PLAYING since a flush event
    in the PAUSED state could unqueue the buffer again.