Revert "mmalvideosink: Fix buffer unref in case of error in show_frame()"
This change could lead to double-free: buffers are unreferenced in the base
sink class even if the result returned from `show_frame()` indicates an
error.  `GstVideoSink` does not need to care about incoming buffer lifetime
in any case

Besides, this logic was more an assertion where it no longer matters if
buffers are leaked or not as continuing execution after this statement
fails renders undefined behaviour.

This reverts commit 067abd0ebe59898ce6e83db46a7c42678559af47.