mmaldeinterlace: Set component frame duration based on frame rate
Setting frame duration is essential for deinterlace component.  It allows
it to work out parameters for compensating unwanted artefacts.  In
particular some flicker was noticed on one of the test streams [1].  It was
likely to do with incorrect frame rate (duration) assumed by default so
setting it explicitly improves it evidently.

Also deinterlacing doubles the frame rate on the output so inditacting this
on the output caps/parameters as well.

[1] http://rdmedia.bbc.co.uk/dash/ondemand/testcard/1/client_manifest-events.mpd