mmalglupload: Add new element to convert MMAL opaque buffers to GLMemory buffers.
This commit adds a new `mmalglupload` element which takes MMAL opaque
buffers (such as produced by `mmalh264dec`, for instance), and produces
GLMemory buffers that can be fed straight into `glimagesinkelement`.

Example usage:

  GST_GL_WINDOW=dispmanx gst-launch-1.0 uridecodebin \
  uri="http://rdmedia.bbc.co.uk/dash/ondemand/testcard/1/client_manifest-720p-stereo-events.mpd" \
  caps='video/x-h264; audio/x-raw' name=d d. ! queue ! h264parse ! \
  mmalh264dec ! mmalglupload ! glimagesinkelement d. ! 'audio/x-raw' ! \
  audioconvert ! alsasink

NOTE: This element is rather experimental at present, and the implementation
contains some hacks.  It also requires support for "external-oes" texture
target, among other things, so will only be built if gststreamer-gl >= 1.7.2
is available.

Playback is generally not as smooth as with `mmalvideosink`, so if you
just want to play video, you're better off using that.

Also, it seems that `glimagesinkelement` can't scale the video on it's
own, so another element is required to do that.  Unfortunately,
`glcolorscale` doesn't support the "external-oes" texture target
required by `mmalglupload`, so that cannot be used.

Implementation notes:

- The implementation creates EGLImage handles to the video data pointed
to by the MMAL opaque buffer pointers.  It can be quite expensive to
destroy the EGLImage instances when we're done with them (measured over
20ms), so we keep a cache of MMAL opaque buffer ptr -> corresponding
EGLImage instance.  There is a fixed-size pool of MMAL buffers held
upstream, so the size of the cache is bounded.  When the MMAL buffers
are repopulated with new data the EGLImage automatically reflects that.

- We use a GL buffer pool provided by downstream to allocate output
buffers.  Just prior to rendering we make the texture held by the
GLMemory a sibling of the EGLImage with a call to
`glEGLImageTargetTexture2DOES()`, using the `GL_TEXTURE_EXTERNAL_OES`
target.  We currently achieve this via `GstGLSyncMeta`, which avoids
us having to block our upstream thread.  Posting the call to
`glEGLImageTargetTexture2DOES()` onto the GL thread and waiting can
take up to 70ms or so to complete, and that is unacceptable.

- If `GstGLUploadMeta` allowed us to express the texture target (as
"external-oes", and `glupload` respected the target set by the meta,
then there's a good chance that we could avoid the need for this upload
element altogether, and instead just add the meta to the buffers output
by our decoder and deinterlacer elements, and include the required
upload functionality in our "MMAL opaque" GstMemory implementation.