mmalclock: Add mmalclock
MMAL uses VideoCore clock internally to synchronise media.  Although it
is not a clock that is used for controlling video output, it has much
better chance to be more accurate.

This commit provides implementation of `GstSystemClock` based on MMAL
clock with the intent to be exposed as the clock that controls GStreamer
pipeline.