mmalvideosink: Do not set reference clock unless connected
When a clock port is set as a reference, it attempts to forward events
(time value and parameter changes) to other clock ports.  This
presumably requires extra buffers allocated which is the case when the
port is connected to any other port.

In 'mmalvideosink' scheduler clock port is not connected to any other
port as currently this is the only MMAL component requiring clock.
Setting it as a reference clock results in ENOSPC errors which is
failing to forward events but the media time is still set.

To avoid spurious errors, this commit removes code that sets scheduler
clock port as a reference clock.