mmalvideodec: Improve acquiring input buffers and handling flush
'mmalvideodec' was occasionally reporting a failure to acquire input
buffer after a 2 seconds time out, especially when transitioning from
low resolution and low frame rate representation to high resolution and
high frame rate.  This could be either down to low frame rate MMAL
buffers still queued in the scheduler and waiting for their turn (for
example it takes 3.2 seconds to render 20 frames at 25/4 frame rate) or
the VC decoder being busy and not producing output to be consumed
quickly enough after representation change.

A mitigation is to wait indefinitely for input buffer which worked fine
but would block the stream thread when flushing (seeking or stopping).
The previous timed wait approach would also have similar problem as it'd
take up to 2 seconds to acquire a buffer and return from the
`handle_frame()` function.

The solution is to use `FLUSH_START` and `FLUSH_STOP` events.  In the
handler for the former one, MMAL input port is flushed which unblocks
the function that waits for an input buffer without additional
synchronisation required.  In `FLUSH_STOP` (where
`GstVideoDecode::stop()` is called), all existing behaviour is
preserved.

With this commit, no more failures were seen and also seeking seems a
bit snappier.