mmaldeinterlace: Only disable control port if it's enabled
In some scenarios attempts are made to disable control port which is not
enabled yet or was disabled already (eg. closing the pipeline before
getting into playing state).  This results in an error printed from
MMAL:

  mmal: mmal_port_disable: port vc.ril.image_fx:in:0(0x71760450) is not enabled

By convention, port status should be checked before trying to disable it
and this commit does just that.