mmaldeinterlace: Deactivate and release output pool only if valid
In situations when 'mmaldeinterlace' was stopped before actually had a
chance to configure the output, the following assertion was seen:

  GStreamer-CRITICAL **: gst_buffer_pool_set_active: assertion 'GST_IS_BUFFER_POOL (pool)' failed

This is because at the time 'mmaldeinterlace' was stopped, the output
buffer pull was still `NULL`.  An example of such situation is failing
to negotiate caps ('mmaldeinterlace' supports only MMAL opaque buffers).

This commit ensures that the buffer pool is accessed and unreferenced
only when valid.