mmal: Fixate the number of opaque buffers and their size
MMAL opaque buffer number and size was adapted to the actual media
format and according to recommendations reported on MMAL ports.  This
meant that for adaptive streams the format and buffer requirements could
change during the playback.  MMAL doesn't support very well changing
this type of configuration dynamically and ideally media format should
be decided only once.  Changing it during the playback worked most of
the time though.

With the intent to introduce MMAL scheduler in 'mmalvideosink', this
scheme can no longer work.  The reason is that in opaque mode buffer
pools are associated with ports and after ports are reconfigured, buffer
pools need to be reconfigured as well (mainly the buffer size).  This
means destroying buffer pools and creating them with new parameters [1].

If any downstream element has a queue of MMAL buffers (like MMAL
scheduler), this can cause havoc as some buffers may still be in use
(for example queued for rendering) and when released, their buffer pool
might not exist any more.  On the other hand, deferring buffer pool
destruction would require keeping a reference to a port in order to
destroy it properly [2] but the port configuration would already be
changed potentially breaking the consistency.

As a compromise, this commit fixates the number and the size of opaque
buffers between MMAL elements.  The maximum supported 1920x1080
resolution is used to calculate the fixated buffer size and the number
of buffer headers is picked to be 20 (a rather arbitrary number of
buffers already used by decode and deinterlace elements).  This allows
to create buffer pools only once which avoids any buffer reconfiguration
during the playback.

[1] `mmal_pool_resize()` is to no avail as unfortunately it destroys all
    buffer headers that might still be in use elsewhere, ignoring any
    reference counting.

[2] See `mmal_port_pool_destroy()`.  Opaque buffers can be allocated
    only in a pool associated with a port.