mmalvideosink: Provide clock to the pipeline
Each element in GStreamer pipeline can provide a clock if it has one.
'mmalvideosink' uses MMAL clock and with this commit provides it to the
pipeline as a better alternative to the default system clock.  If any
other element's clock is used instead (for example audio sink clock),
this still works fine as MMAL clock framework compensates any jitter.