mmalvideodec,mmaldeinterlace: Do not disable MMAL ports when seeking
For some streams MMAL decoder was ending up in a state after seek when
it accepted input buffers and returned them but did not produce any
output buffers.  Also pausing did not work for some DASH streams if not
sought in the stream first.

The problem seemed to be related to flush operation implemented in
GStreamer components which in addition to flushing MMAL ports also
disabled them.  Disabling MMAL port is a rather radical operation and it
seemed that it should be avoided during the playback.

With this commit MMAL ports are not disabled when flushing during the
playback (eg. after seeking) which makes seeking and pausing work in all
cases.  This change is also made is 'mmaldeinterlace' as it uses similar
concepts and implementation.