mmalvideodec: Improve logging in MMAL callbacks and elsewhere
MMAL decoder had no logging in MMAL input/output port callbacks which
made it difficult to reason about MMAL buffer use and data flow.  This
commit adds logging (TRACE level) to show when an input buffer is
returned to the pool and when decoded data is queued for sending
downstream.

Also minor additions and changes are made elsewhere to better understand
the data flow and improve readability (time format).