mmalvideodec: Handle seeking while changing input format
Seek requests are usually made by the application independently of the
stream flow.  This means that the elements may be performing any
(potentially blocking) operation during the request.  For that reason
FLUSH_START is supposed to unblock the stream in each element.

Specifically, MMAL decoder blocks waiting on MMAL input buffer to drain
the decoder component before applying new configuration on format change
(adaptive streams). This is cancelled by FLUSH_START.  The problem is
that the format change is not completed and if after seek new format is
played, the previous configuration is still in use and as a result the
decoder component usually doesn't produce any output.

This commit mitigates the problem by deferring format change until
FLUSH_STOP (synchronised with the stream) when the element should become
ready to accept new buffers.  After this change seeking from any
representation of adaptive stream works [1].

[1] For example seek (right arrow on the keyboard) on the first
    representation:

    gst-play-1.0 --audiosink alsasink \
        --videosink 'h264parse ! mmalh264dec ! mmaldeinterlace ! mmalvideosink' \
        http://rdmedia.bbc.co.uk/dash/ondemand/testcard/1/client_manifest-events.mpd