mmalvideosink: Disable input port if format change not supported
After adding MMAL scheduler to 'mmalvideosink' and fixating buffer
pools, switching between opaque and plain buffers does not work any
more.  This requires more investigation but for the time being an error
is raised and the pipeline is stopped if switching between opaque and
plain buffers is requested.

Note that using plain buffers itself works, it's just switching between
opaque and plain is not supported.