mmalvideosink: Double the number of declared opaque input buffers
In opaque mode MMAL buffers are allocated upstream.  Each component
needs to know the maximum number of buffers it should expect on its
input to allocate internal data.  If more buffers than `buffer_num` are
sent, ENOMEM is returned.

There can be MMAL elements upstream that work in pass-through mode and
do not allocate/send MMAL buffers.  But if they start processing data,
their buffers are received downstream.  This on its own is not a problem
as those elements need the other upstream MMAL buffers to produce their
data.  But elements like 'mmaldeinterlace' double the frame rate and send
twice as much buffers as on their input.  We can end up in a situation
when there's `GST_MMAL_NUM_OUTPUT_BUFFERS - 1` buffers queued in the
scheduler, 'mmaldeinterlace' goes out of pass-through mode after a
format change and from the remaining buffer it produces two buffers
which results in `GST_MMAL_NUM_OUTPUT_BUFFERS + 1` total buffers sent to
the sink (scheduler input port) which then causes ENOMEM.

For the reasons described above, this commit makes room for twice as
much buffers in the scheduler as in the other components.  These are
actually buffer _headers_ which are cheap to allocate (128 bytes) so
memory increase should be negligible.