mmalvideodec: Support plain and opaque buffer with fixated pools
MMAL decoder is one of the elements that has an output pool of opaque
buffers with the buffers referenced and sent downstream [1].  Before
fixating buffer pools with one of the previous commits, it could switch
between opaque (preferred) and plain buffers dynamically by simply
destroying the pool and recreating it with new configuration.  This
could be required in cases when elements not supporting MMAL opaque
buffers were present downstream in the pipeline and switched its
pass-through mode (for example based on current media format) [2].

Since output buffer pools are now fixated and MMAL opaque buffers are
preferred, this commit creates a fixated MMAL plain buffer pool [3] only
when needed.  The decoder then switches between the two.

[1] MMAL deinterlace element is another example but doesn't support
    plain buffers.

[2] An example of such element is GStreamer 'deinterlace' which is
    "activated" in case of adaptive stream changing its representation
    to interlaced.

[3] In case of plain buffers the pool does not need to be fixated.  For
    simplicity and symmetry with the opaque buffer pool, it's fixated as
    well.